Combined align and tiling stacks. A stack deployment now has a unique resource locaiton in S3.
Alignment and tiling stacks are now nested stacks under
console.template, which has the front-end. Nesting should make it easier
to develop a new stack for a new type of job separately, then hook it up
to the same console and deploy it along with the rest of the landform
pipeline.

Drawback: the deployment scripts grow in complexity when new nested
stacks and lambdas are added.